GATEWAY-3051: unit tests
